1. Updated transaction usage in DbUpdate: Since ISqlExecuter now always uses the scope-level IPersistenceTransaction for each command, DbUpdate scope needs to have PersistenceTransaction registered and configured to not create a database transaction. 2. TRANCOUNT verification is moved from MsSqlExecuter to SqlTransactionBatches. There is low risk of transaction issues at runtime with MsSqlExecuter so there is no need to hinder performance. At deploy-time, SqlTransactionBatches is used for DDL SQL scripts where transaction issues present higher risk, and verification after a batch of scripts has low performance impact. 3. ISqlExecuter CheckTransactionCount replaced with GetTransactionCount; its ExecuteScalar is no slower then ExecuteNonQuery, the new method even has some minor performance benefits. 4. SqlTransactionBatches is interfaced with ISqlTransactionBatches.

+++ CommonConcepts/CommonConcepts.Test/FrameworkIntegrationTests/MsSqlExecuterIntegrationTest.cs
@@ -27,27 +27,56 @@ using System.Collections.Generic;
 using System.Data.Common;
 using System.Data.SqlClient;
 using System.Linq;
+using System.Reflection;
+using System.Text.RegularExpressions;
 
 namespace Rhetos.Persistence.Test
 {
     [TestClass]
     public class MsSqlExecuterIntegrationTest
     {
-        private static MsSqlExecuter NewSqlExecuter(string connectionString = null, IUserInfo testUser = null)
+        private static string LogActionName => MethodBase.GetCurrentMethod().DeclaringType.Name;
+
+        /// <summary>
+        /// Executes the MsSqlExecuter commands and commits the transaction by default.
+        /// </summary>
+        private static void InTransaction(Action<MsSqlExecuter> sqlExecuterAction, string connectionString = null, IUserInfo testUser = null, bool commit = true)
         {
-            connectionString = connectionString ?? SqlUtility.ConnectionString;
-            testUser = testUser ?? new NullUserInfo();
-            return new MsSqlExecuter(connectionString, new ConsoleLogProvider(), testUser, null);
+            connectionString ??= SqlUtility.ConnectionString;
+            testUser ??= new NullUserInfo();
+
+            using (var persistenceTransaction = new PersistenceTransaction(
+                new ConsoleLogProvider(),
+                connectionString,
+                testUser,
+                new PersistenceTransactionOptions()))
+            {
+                var sqlExecuter = new MsSqlExecuter(new ConsoleLogProvider(), persistenceTransaction);
+                sqlExecuterAction.Invoke(sqlExecuter);
+                if (commit)
+                    persistenceTransaction.CommitAndClose();
+            }
         }
 
-        private static MsSqlExecuter NewSqlExecuter(IUserInfo testUser)
+        /// <summary>
+        /// Executes the MsSqlExecuter command and commits the transaction by default.
+        /// </summary>
+        private static void ExecuteSql(IEnumerable<string> commands, string connectionString = null, IUserInfo testUser = null, bool commit = true)
         {
-            return NewSqlExecuter(null, testUser);
+            InTransaction(sqlExecuter => sqlExecuter.ExecuteSql(commands), connectionString, testUser, commit);
+        }
+
+        /// <summary>
+        /// Executes the MsSqlExecuter command and commits the transaction by default.
+        /// </summary>
+        private static void ExecuteReader(string command, Action<DbDataReader> action, string connectionString = null, IUserInfo testUser = null, bool commit = true)
+        {
+            InTransaction(sqlExecuter => sqlExecuter.ExecuteReader(command, action), connectionString, testUser, commit);
         }
 
         private string GetRandomTableName()
         {
-            NewSqlExecuter().ExecuteSql(new[] { "IF SCHEMA_ID('RhetosUnitTest') IS NULL EXEC('CREATE SCHEMA RhetosUnitTest')" });
+            ExecuteSql(new[] { "IF SCHEMA_ID('RhetosUnitTest') IS NULL EXEC('CREATE SCHEMA RhetosUnitTest')" });
             var newTableName = "RhetosUnitTest.T" + Guid.NewGuid().ToString().Replace("-", "");
             Console.WriteLine("Generated random table name: " + newTableName);
             return newTableName;
@@ -66,24 +95,29 @@ namespace Rhetos.Persistence.Test
         [ClassCleanup]
         public static void DropRhetosUnitTestSchema()
         {
-            try { TestUtility.CheckDatabaseAvailability("MsSql"); }
-            catch { return; }
-
             Console.WriteLine("=== ClassCleanup ===");
 
-            NewSqlExecuter().ExecuteSql(new[] { 
-                @"DECLARE @sql NVARCHAR(MAX)
-                    SET @sql = ''
-                    SELECT @sql = @sql + 'DROP TABLE RhetosUnitTest.' + QUOTENAME(name) + ';' + CHAR(13) + CHAR(10)
-                        FROM sys.tables WHERE schema_id = SCHEMA_ID('RhetosUnitTest')
-                    EXEC (@sql)",
-                "IF SCHEMA_ID('RhetosUnitTest') IS NOT NULL DROP SCHEMA RhetosUnitTest" });
+            using (var scope = TestScope.Create())
+            {
+                var sqlExecuter = scope.Resolve<ISqlExecuter>();
+                sqlExecuter.ExecuteSqlInterpolated(
+                    $"DELETE FROM Common.Log WHERE Action = {LogActionName} AND TableName IS NULL");
+                scope.CommitAndClose();
+            }
+
+            if (SqlUtility.DatabaseLanguage == "MsSql")
+                ExecuteSql(new[] {
+                    @"DECLARE @sql NVARCHAR(MAX)
+                        SET @sql = ''
+                        SELECT @sql = @sql + 'DROP TABLE RhetosUnitTest.' + QUOTENAME(name) + ';' + CHAR(13) + CHAR(10)
+                            FROM sys.tables WHERE schema_id = SCHEMA_ID('RhetosUnitTest')
+                        EXEC (@sql)",
+                    "IF SCHEMA_ID('RhetosUnitTest') IS NOT NULL DROP SCHEMA RhetosUnitTest" });
         }
 
         [TestMethod]
         public void ExecuteSql_SaveLoadTest()
         {
-            MsSqlExecuter sqlExecuter = NewSqlExecuter();
             string table = GetRandomTableName();
 
             IEnumerable<string> commands = new[]
@@ -91,30 +125,28 @@ namespace Rhetos.Persistence.Test
                     "CREATE TABLE " + table + " ( A INTEGER )",
                     "INSERT INTO " + table + " SELECT 123"
                 };
-            sqlExecuter.ExecuteSql(commands);
+            ExecuteSql(commands);
             int actual = 0;
-            sqlExecuter.ExecuteReader("SELECT * FROM " + table, dr => actual = dr.GetInt32(0));
+            ExecuteReader("SELECT * FROM " + table, dr => actual = dr.GetInt32(0));
             Assert.AreEqual(123, actual);
         }
 
         [TestMethod]
         public void ExecuteSql_SimpleSqlError()
         {
-            TestUtility.ShouldFail(() => NewSqlExecuter().ExecuteSql(new[] { "raiserror('aaa', 16, 100)" }),
+            TestUtility.ShouldFail(() => ExecuteSql(new[] { "raiserror('aaa', 16, 100)" }),
                 "aaa", "16", "100");
         }
 
         [TestMethod]
         public void ExecuteSql_InfoMessageIsNotError()
         {
-            NewSqlExecuter().ExecuteSql(new[] { "raiserror('aaa', 0, 100)" }); // Exception not expected here.
+            ExecuteSql(new[] { "raiserror('aaa', 0, 100)" }); // Exception not expected here.
         }
 
         [TestMethod]
         public void ExecuteSql_ErrorDescriptions()
         {
-            MsSqlExecuter sqlExecuter = NewSqlExecuter();
-              
             IEnumerable<string> commands = new[]
                 {
 @"print 'xxx'
@@ -128,7 +160,7 @@ raiserror('fff', 18, 118)"
 
             var expectedStrings = new[] { "aaa", "bbb", "ccc", "ddd", "eee", "fff", "101", "116", "117", "118" }; // Error of severity 0 and 10 (states "100" and "110") do not require detailed error message.
 
-            TestUtility.ShouldFail(() => sqlExecuter.ExecuteSql(commands), expectedStrings);
+            TestUtility.ShouldFail(() => ExecuteSql(commands), expectedStrings);
         }
 
         [TestMethod]
@@ -143,8 +175,7 @@ raiserror('fff', 18, 118)"
             string nonexistentDatabaseConnectionString = connectionStringBuilder.ConnectionString;
             Console.WriteLine(nonexistentDatabaseConnectionString);
 
-            MsSqlExecuter sqlExecuter = NewSqlExecuter(nonexistentDatabaseConnectionString);
-            TestUtility.ShouldFail(() => sqlExecuter.ExecuteSql(new[] { "print 123" }),
+            TestUtility.ShouldFail(() => ExecuteSql(new[] { "print 123" }, nonexistentDatabaseConnectionString),
                 connectionStringBuilder.DataSource, connectionStringBuilder.InitialCatalog, Environment.UserName);
         }
 
@@ -152,9 +183,9 @@ raiserror('fff', 18, 118)"
         public void ExecuteSql_CommitImmediately()
         {
             string table = GetRandomTableName();
-            NewSqlExecuter().ExecuteSql(new[] {
+            ExecuteSql(new[] {
                 "create table " + table + " ( a integer )" });
-            NewSqlExecuter().ExecuteSql(new[] {
+            ExecuteSql(new[] {
                 "set lock_timeout 0",
                 "select * from " + table }); // Exception not expected here.
         }
@@ -165,14 +196,14 @@ raiserror('fff', 18, 118)"
             string table = GetRandomTableName();
             try
             {
-                NewSqlExecuter().ExecuteSql(new[] {
+                ExecuteSql(new[] {
                     "create table " + table + " ( a integer )",
                     "create table forcederror" });
             }
             catch
             {
             }
-            NewSqlExecuter().ExecuteSql(new[] {
+            ExecuteSql(new[] {
                 "set lock_timeout 0",
                 "create table " + table + " ( a integer )" }); // Lock timeout exception not expected here.
         }
@@ -183,46 +214,91 @@ raiserror('fff', 18, 118)"
             string table = GetRandomTableName();
             try
             {
-                NewSqlExecuter().ExecuteSql(new[] {
+                ExecuteSql(new[] {
                     "create table forcederror",
                     "create table " + table + " ( a integer )" });
             }
             catch
             {
             }
-            NewSqlExecuter().ExecuteSql(new[] {
+            ExecuteSql(new[] {
                 "create table " + table + " ( a integer )" }); // Exception not expected here.
         }
 
         [TestMethod]
         public void ExecuteSql_RollbackedTransaction()
         {
-            TestUtility.ShouldFail<FrameworkException>(
-                () => NewSqlExecuter().ExecuteSql(new[] {
-                    "print '1'",
-                    "rollback",
-                    "print '2'" }),
-                "transaction", "rollback");
+            // Since Rhetos v5 there is no explicit check for transaction level in MsSqlExecuter (only in SqlTransactionBatches),
+            // but the unit of work should still fail on transaction commit at the end of the scope, because it is no longer open.
+
+            Guid id = Guid.NewGuid();
+            TestUtility.ShouldFail<InvalidOperationException>(
+                () => ExecuteSql(
+                    new[] {
+                        $"INSERT INTO Common.Log (Action, ItemId, Description) SELECT '{LogActionName}', '{id}', '1'",
+                        $"ROLLBACK",
+                        $"INSERT INTO Common.Log (Action, ItemId, Description) SELECT '{LogActionName}', '{id}', '2'" }),
+                "This SqlTransaction has completed; it is no longer usable.");
+
+            // Record '2' was unintentionally inserted out of transaction, since MsSqlExecuter does not check the transaction state
+            // after each command, for performance reasons.
+            var result = new List<string>();
+            ExecuteReader(
+                $"SELECT Description FROM Common.Log WHERE ItemId = '{id}'",
+                reader => result.Add(reader.GetString(0)));
+            Assert.AreEqual("2", TestUtility.DumpSorted(result));
         }
 
         [TestMethod]
-        public void ExecuteSql_TransactionLevel()
+        public void ExecuteSql_RollbackedTransactionCanceled()
         {
-            TestUtility.ShouldFail<FrameworkException>(
-                () => NewSqlExecuter().ExecuteSql(new[] {
-                    "print '1'",
-                    "begin tran",
-                    "print '2'" }),
-                "transaction", "begin tran");
+            // Since Rhetos v5 there is no explicit check for transaction level in MsSqlExecuter (only in SqlTransactionBatches),
+            // but the unit of work should still fail on transaction commit at the end of the scope, because it is no longer open.
+
+            Guid id = Guid.NewGuid();
+            ExecuteSql(
+                new[] {
+                    $"INSERT INTO Common.Log (Action, ItemId, Description) SELECT '{LogActionName}', '{id}', '1'",
+                    $"ROLLBACK",
+                    $"INSERT INTO Common.Log (Action, ItemId, Description) SELECT '{LogActionName}', '{id}', '2'" },
+                commit: false);
+
+            // Record '2' was unintentionally inserted out of transaction, since MsSqlExecuter does not check the transaction state
+            // after each command, for performance reasons.
+            var result = new List<string>();
+            ExecuteReader(
+                $"SELECT Description FROM Common.Log WHERE ItemId = '{id}'",
+                reader => result.Add(reader.GetString(0)));
+            Assert.AreEqual("2", TestUtility.DumpSorted(result));
+        }
+
+        [TestMethod]
+        public void ExecuteSql_TransactionNotClosed()
+        {
+            Guid id = Guid.NewGuid();
+            ExecuteSql(
+                new[] {
+                    $"INSERT INTO Common.Log (Action, ItemId, Description) SELECT '{LogActionName}', '{id}', '1'",
+                    $"BEGIN TRAN",
+                    $"INSERT INTO Common.Log (Action, ItemId, Description) SELECT '{LogActionName}', '{id}', '2'" });
+
+            // No records have been committed (even though there was no error), since MsSqlExecuter does not check the transaction state
+            // after each command, for performance reasons.
+            // After "BEGIN TRAN", the trancount have been increased to 2, PersistenceTransaction's commit
+            // just reduced the trancount to 1, and the SqlTransaction/SqlConnection disposal rolled back the transaction.
+            var result = new List<string>();
+            ExecuteReader(
+                $"SELECT Description FROM Common.Log WHERE ItemId = '{id}'",
+                reader => result.Add(reader.GetString(0)));
+            Assert.AreEqual("", TestUtility.DumpSorted(result));
         }
 
         [TestMethod]
         public void SendUserInfoInSqlContext_NoUser()
         {
             var testUser = new TestUserInfo(null, null, false);
-            var sqlExecuter = NewSqlExecuter(SqlUtility.ConnectionString, testUser);
             var result = new List<object>();
-            sqlExecuter.ExecuteReader("SELECT context_info()", reader => result.Add(reader[0]));
+            ExecuteReader("SELECT context_info()", reader => result.Add(reader[0]), SqlUtility.ConnectionString, testUser);
 
             Console.WriteLine(result.Single());
             Assert.AreEqual(typeof(DBNull), result.Single().GetType());
@@ -242,14 +318,13 @@ raiserror('fff', 18, 118)"
         public void SendUserInfoInSqlContext_WriteWithUser()
         {
             var testUser = new TestUserInfo("Bob", "HAL9000");
-            var sqlExecuter = NewSqlExecuter(testUser);
             string table = GetRandomTableName();
 
             var result = new List<string>();
-            sqlExecuter.ExecuteSql($"SELECT Context = {contextInfoToText} INTO {table}");
-            sqlExecuter.ExecuteReader(
+            ExecuteSql(new[] { $"SELECT Context = {contextInfoToText} INTO {table}" }, testUser: testUser);
+            ExecuteReader(
                 @"SELECT * FROM " + table,
-                reader => result.Add(reader[0].ToString()));
+                reader => result.Add(reader[0].ToString()), testUser: testUser);
 
             Assert.AreEqual("Rhetos:Bob,HAL9000", TestUtility.Dump(result));
         }
@@ -275,14 +350,13 @@ raiserror('fff', 18, 118)"
 
         private static string ReadContextWithSqlExecuter(TestUserInfo testUser)
         {
-            var sqlExecuter = NewSqlExecuter(testUser);
             var result = new List<string>();
-            sqlExecuter.ExecuteReader(
+            ExecuteReader(
                 $"SELECT {contextInfoToText}",
-                reader => result.Add(reader[0].ToString()));
+                reader => result.Add(reader[0].ToString()),
+                testUser: testUser);
             return result.Single();
         }
-
 
         [TestMethod]
         public void ConsistentRead()
@@ -325,8 +399,8 @@ raiserror('fff', 18, 118)"
             }
 
             Assert.AreEqual(
-                string.Join(Environment.NewLine, msSqlExecuterReport),
-                string.Join(Environment.NewLine, baseSqlExecuterReport),
+                CleanupSqlExecuterLog(string.Join(Environment.NewLine, msSqlExecuterReport)),
+                CleanupSqlExecuterLog(string.Join(Environment.NewLine, baseSqlExecuterReport)),
                 $"{nameof(msSqlExecuterReport)} result does not match {nameof(baseSqlExecuterReport)}.");
         }
 
@@ -357,6 +431,25 @@ raiserror('fff', 18, 118)"
             lastSizeAfterReader = lastSizeAfterReader ?? log.Count;
             return "Result: " + queryResult + Environment.NewLine + "Log:" + Environment.NewLine + string.Join(Environment.NewLine, log.Take(lastSizeAfterReader.Value));
         }
+
+        private static string CleanupSqlExecuterLog(string report)
+        {
+            // Performance log includes milliseconds that are not reproducible.
+            report = subSecondDigitsRegex.Replace(report, "*ms");
+
+            // MsSqlExecuter's method supports multiple SQL commands, hence the additional information in log.
+            report = report.Replace("[Trace] MsSqlExecuter: Executing 1 commands.\r\n", "");
+
+            // MsSqlExecuter's method does not support parameters.
+            report = sqlStringRegex.Replace(report, "*param");
+            report = report.Replace("{0}", "*param");
+
+            return report;
+        }
+
+        private static Regex subSecondDigitsRegex = new Regex(@"\b\d{7}\b");
+
+        private static Regex sqlStringRegex = new Regex(@"'.*?'");
 
         [TestMethod]
         public void ConsistentExecute()
@@ -399,8 +492,8 @@ raiserror('fff', 18, 118)"
             }
 
             Assert.AreEqual(
-                string.Join(Environment.NewLine, msSqlExecuterReport),
-                string.Join(Environment.NewLine, baseSqlExecuterReport),
+                CleanupSqlExecuterLog(string.Join(Environment.NewLine, msSqlExecuterReport)),
+                CleanupSqlExecuterLog(string.Join(Environment.NewLine, baseSqlExecuterReport)),
                 $"{nameof(msSqlExecuterReport)} result does not match {nameof(baseSqlExecuterReport)}.");
         }
